requirements.txt
pyrogram
python-dotenv
tgcrypto